change my code

+++ src/main/java/org/example/free_new_magazine/security/CustomUserDetails.java
@@ -37,5 +37,5 @@ public class CustomUserDetails implements UserDetails {
     @Override
     public boolean isEnabled() { return true; }
 
-    public Long getId(){ return user.getId(); }
+    public Integer getId(){ return user.getId(); }
 }
